@@ -1,8 +1,8 @@
 function Validation(values){
     
     let error = {}
-    const email_pattern  = /^[^\s@]+@[^\s@]+\. [^\s@]+$/
-    const password_pattern =  /^(?=.*\d) (?=.*[a-z])(?=.*[A-Z])[a-zA-Z0-9]{8,}$/
+    const email_pattern  = /^[^\s@]+@[^\s@]+\.[^\s@]+$/
+    const password_pattern =  /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])[a-zA-Z0-9]{8,}$/
 
     if (values.email === "") {
     error.email = "Name should not be empty"
