@@ -1,52 +1,395 @@
-import {Link} from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import React, { useState } from 'react'
-import Validation from './SignupValidation'  
-
+import Validation from './SignupValidation'
+import { motion, AnimatePresence } from 'framer-motion';
 
 function Signup() {
-    const[values , setValues] = useState({
-        name:'',
-        email:'',
-        password:''
-      })
-      const [errors , setErrors] = useState({})
-    const handleInput = (event) => {
-        setValues({...values,[event.target.name]:event.target.value})
+  const [values, setValues] = useState({
+    name: '',
+    email: '',
+    password: '',
+    confirmPassword: ''
+  })
+  const [errors, setErrors] = useState({})
+  const [isLoading, setIsLoading] = useState(false)
+  const [step, setStep] = useState(1) // Multi-step form
+  const [passwordStrength, setPasswordStrength] = useState(0)
+
+  const handleInput = (event) => {
+    setValues({ ...values, [event.target.name]: event.target.value })
+    
+    // Calculate password strength
+    if (event.target.name === 'password') {
+      calculatePasswordStrength(event.target.value)
     }
-      const handleSubmit = (event) => {
-        event.preventDefault()
-        setErrors(Validation(values));
+  }
+
+  const calculatePasswordStrength = (password) => {
+    let strength = 0
+    
+    // Check password length
+    if (password.length >= 8) strength += 25
+    
+    // Check for uppercase letters
+    if (password.match(/[A-Z]/)) strength += 25
+    
+    // Check for numbers
+    if (password.match(/[0-9]/)) strength += 25
+    
+    // Check for special characters
+    if (password.match(/[^a-zA-Z0-9]/)) strength += 25
+    
+    setPasswordStrength(strength)
+    
+    return {
+      score: strength,
+      isStrong: strength >= 75,
+      isMedium: strength >= 50 && strength < 75,
+      isWeak: strength < 50 && strength > 0,
+      isEmpty: strength === 0
+    }
+  }
+
+  const handleSubmit = (event) => {
+    event.preventDefault()
+    const validationErrors = Validation(values)
+    setErrors(validationErrors)
+    
+    if (Object.keys(validationErrors).length === 0) {
+      setIsLoading(true)
+      
+      // Simulate API call
+      setTimeout(() => {
+        setIsLoading(false)
+        // Here you would handle successful signup
+      }, 1500)
+    }
+  }
+
+  const nextStep = () => {
+    setStep(step + 1)
+  }
+
+  const prevStep = () => {
+    setStep(step - 1)
+  }
+
+  // Animation variants
+  const containerVariants = {
+    hidden: { opacity: 0, y: 20 },
+    visible: { 
+      opacity: 1, 
+      y: 0,
+      transition: { 
+        duration: 0.5,
+        when: "beforeChildren",
+        staggerChildren: 0.1
       }
+    },
+    exit: { 
+      opacity: 0, 
+      y: -20,
+      transition: { duration: 0.3 }
+    }
+  }
+
+  const itemVariants = {
+    hidden: { opacity: 0, x: -10 },
+    visible: { opacity: 1, x: 0 },
+    exit: { opacity: 0, x: 10 }
+  }
 
   return (
-    <div className='d-flex justify-content-center align-items-center bg-primary vh-100'>
-        <div className='bg-white p5 rounded w-50'>
-            <h2>Sign-Up</h2>
-            <form action ="" onSubmit={handleSubmit}>
-                <div className='mb-3'>
-                   <label htmlFor='name'><strong>Name</strong></label>
-                   <input type="text" placeholder='Enter Name' name='email'  
-                   onChange={handleInput} className='form-control rounded-0' required/>
-                   {errors.name && <span className='text-danger'> {errors.name}</span>}
-                 </div>
-                 <div className='mb-3'>
-                   <label htmlFor='email'><strong>Email</strong></label>
-                   <input type="email" placeholder='Enter Email' name='email'  
-                   onChange={handleInput} className='form-control rounded-0' required/>
-                  {errors.email && <span className='text-danger'> {errors.email}</span>}
-                 </div>
-                 <div className='mb=3'>
-                    <label htmlFor='password'><strong>Password</strong></label>
-                    <input type="password" placeholder='Enter password' name='password' 
-                    onChange={handleInput} className='form-control rounded-0'required/>
-                    {errors.password && <span className='text-danger'> {errors.password}</span>}
-                 </div>
-                 <button type = 'submit' className='btn btn-success w-100 rounded-0'><strong>SIgn up</strong></button>
-                 <p> You agree to our terms and conditions</p>
-                 <Link to ="/"className='btn btn-default border w-100 bg-light rounded-0 text-decoration-none'>Login</Link>
-            </form>
+    <motion.div 
+      variants={containerVariants}
+      initial="hidden"
+      animate="visible"
+      exit="exit"
+      className="w-full max-w-md relative z-10"
+    >
+      {/* Decorative elements */}
+      <div className="absolute -top-10 -right-10 w-32 h-32 bg-indigo-500 rounded-full mix-blend-multiply filter blur-xl opacity-20 animate-float"></div>
+      <div className="absolute -bottom-10 -left-10 w-32 h-32 bg-purple-500 rounded-full mix-blend-multiply filter blur-xl opacity-20 animate-float" style={{ animationDelay: "1s" }}></div>
+      
+      <div className="bg-white shadow-2xl rounded-3xl overflow-hidden backdrop-blur-lg bg-opacity-95 border border-gray-100 relative z-10">
+        <div className="px-10 py-8">
+          <motion.div variants={itemVariants} className="text-center mb-8">
+            <h2 className="text-3xl font-bold mb-2 text-gray-800 font-display">Create Account</h2>
+            <p className="text-gray-500">Join our community of creative minds</p>
+          </motion.div>
+          
+          {/* Progress steps */}
+          <motion.div variants={itemVariants} className="flex justify-center mb-8">
+            <div className="flex items-center">
+              <div className={`flex items-center justify-center w-8 h-8 rounded-full ${step >= 1 ? 'bg-indigo-600 text-white' : 'bg-gray-200 text-gray-600'}`}>
+                1
+              </div>
+              <div className={`w-12 h-1 ${step >= 2 ? 'bg-indigo-600' : 'bg-gray-200'}`}></div>
+              <div className={`flex items-center justify-center w-8 h-8 rounded-full ${step >= 2 ? 'bg-indigo-600 text-white' : 'bg-gray-200 text-gray-600'}`}>
+                2
+              </div>
+              <div className={`w-12 h-1 ${step >= 3 ? 'bg-indigo-600' : 'bg-gray-200'}`}></div>
+              <div className={`flex items-center justify-center w-8 h-8 rounded-full ${step >= 3 ? 'bg-indigo-600 text-white' : 'bg-gray-200 text-gray-600'}`}>
+                3
+              </div>
+            </div>
+          </motion.div>
+          
+          <form onSubmit={handleSubmit} className="space-y-6">
+            <AnimatePresence mode="wait">
+              {step === 1 && (
+                <motion.div 
+                  key="step1"
+                  initial={{ opacity: 0, x: 20 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -20 }}
+                  transition={{ duration: 0.3 }}
+                  className="space-y-6"
+                >
+                  <motion.div variants={itemVariants}>
+                    <label htmlFor="name" className="block text-sm font-medium text-gray-700 mb-1">Full Name</label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
+                        </svg>
+                      </div>
+                      <motion.input
+                        type="text"
+                        id="name"
+                        name="name"
+                        placeholder="John Doe"
+                        value={values.name}
+                        onChange={handleInput}
+                        whileFocus={{ scale: 1.01 }}
+                        transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                        className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                        required
+                      />
+                    </div>
+                    {errors.name && <p className="mt-1 text-sm text-red-600">{errors.name}</p>}
+                  </motion.div>
+                  
+                  <motion.div variants={itemVariants}>
+                    <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">Email Address</label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207" />
+                        </svg>
+                      </div>
+                      <motion.input
+                        type="email"
+                        id="email"
+                        name="email"
+                        placeholder="you@example.com"
+                        value={values.email}
+                        onChange={handleInput}
+                        whileFocus={{ scale: 1.01 }}
+                        transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                        className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                        required
+                      />
+                    </div>
+                    {errors.email && <p className="mt-1 text-sm text-red-600">{errors.email}</p>}
+                  </motion.div>
+
+                  <motion.button
+                    type="button"
+                    onClick={nextStep}
+                    whileHover={{ scale: 1.03 }}
+                    whileTap={{ scale: 0.97 }}
+                    className="w-full py-3 px-4 bg-gradient-to-r from-indigo-600 to-purple-600 text-white font-medium rounded-lg shadow-md hover:shadow-lg transition-all duration-300 ease-in-out"
+                  >
+                    Continue
+                  </motion.button>
+                </motion.div>
+              )}
+
+              {step === 2 && (
+                <motion.div 
+                  key="step2"
+                  initial={{ opacity: 0, x: 20 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -20 }}
+                  transition={{ duration: 0.3 }}
+                  className="space-y-6"
+                >
+                  <motion.div variants={itemVariants}>
+                    <label htmlFor="password" className="block text-sm font-medium text-gray-700 mb-1">Password</label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
+                        </svg>
+                      </div>
+                      <motion.input
+                        type="password"
+                        id="password"
+                        name="password"
+                        placeholder="••••••••"
+                        value={values.password}
+                        onChange={handleInput}
+                        whileFocus={{ scale: 1.01 }}
+                        transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                        className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                        required
+                      />
+                    </div>
+                    
+                    {/* Password strength meter */}
+                    <div className="mt-2">
+                      <div className="flex justify-between mb-1">
+                        <span className="text-xs text-gray-500">Password strength</span>
+                        <span className="text-xs font-semibold">
+                          {passwordStrength <= 25 && "Weak"}
+                          {passwordStrength > 25 && passwordStrength <= 50 && "Fair"}
+                          {passwordStrength > 50 && passwordStrength <= 75 && "Good"}
+                          {passwordStrength > 75 && "Strong"}
+                        </span>
+                      </div>
+                      <div className="w-full bg-gray-200 rounded h-1.5">
+                        <motion.div 
+                          className={`h-1.5 rounded ${
+                            passwordStrength <= 25 ? 'bg-red-500' : 
+                            passwordStrength <= 50 ? 'bg-yellow-500' : 
+                            passwordStrength <= 75 ? 'bg-blue-500' : 'bg-green-500'
+                          }`}
+                          initial={{ width: 0 }}
+                          animate={{ width: `${passwordStrength}%` }}
+                          transition={{ duration: 0.3 }}
+                        ></motion.div>
+                      </div>
+                    </div>
+                    
+                    {errors.password && <p className="mt-1 text-sm text-red-600">{errors.password}</p>}
+                    <p className="text-xs text-gray-500 mt-1">
+                      Password must contain at least 8 characters, one uppercase letter, one lowercase letter, and one number
+                    </p>
+                  </motion.div>
+                  
+                  <motion.div variants={itemVariants}>
+                    <label htmlFor="confirmPassword" className="block text-sm font-medium text-gray-700 mb-1">Confirm Password</label>
+                    <div className="relative">
+                      <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                        <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                          <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M9 12l2 2 4-4m5.618-4.016A11.955 11.955 0 0112 2.944a11.955 11.955 0 01-8.618 3.04A12.02 12.02 0 003 9c0 5.591 3.824 10.29 9 11.622 5.176-1.332 9-6.03 9-11.622 0-1.042-.133-2.052-.382-3.016z" />
+                        </svg>
+                      </div>
+                      <motion.input
+                        type="password"
+                        id="confirmPassword"
+                        name="confirmPassword"
+                        placeholder="••••••••"
+                        value={values.confirmPassword}
+                        onChange={handleInput}
+                        whileFocus={{ scale: 1.01 }}
+                        transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                        className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                        required
+                      />
+                    </div>
+                    {errors.confirmPassword && <p className="mt-1 text-sm text-red-600">{errors.confirmPassword}</p>}
+                  </motion.div>
+
+                  <div className="flex space-x-4">
+                    <motion.button
+                      type="button"
+                      onClick={prevStep}
+                      whileHover={{ scale: 1.03 }}
+                      whileTap={{ scale: 0.97 }}
+                      className="w-1/3 py-3 px-4 bg-gray-200 text-gray-700 font-medium rounded-lg hover:bg-gray-300 transition-all duration-200 ease-in-out"
+                    >
+                      Back
+                    </motion.button>
+                    <motion.button
+                      type="button"
+                      onClick={nextStep}
+                      whileHover={{ scale: 1.03 }}
+                      whileTap={{ scale: 0.97 }}
+                      className="w-2/3 py-3 px-4 bg-gradient-to-r from-indigo-600 to-purple-600 text-white font-medium rounded-lg shadow-md hover:shadow-lg transition-all duration-300 ease-in-out"
+                    >
+                      Continue
+                    </motion.button>
+                  </div>
+                </motion.div>
+              )}
+
+              {step === 3 && (
+                <motion.div 
+                  key="step3"
+                  initial={{ opacity: 0, x: 20 }}
+                  animate={{ opacity: 1, x: 0 }}
+                  exit={{ opacity: 0, x: -20 }}
+                  transition={{ duration: 0.3 }}
+                  className="space-y-6"
+                >
+                  <motion.div variants={itemVariants} className="border border-gray-200 rounded-lg p-4 bg-gray-50">
+                    <h3 className="font-medium mb-2">Account Summary</h3>
+                    <div className="space-y-1 text-sm">
+                      <p><span className="text-gray-500">Name:</span> {values.name}</p>
+                      <p><span className="text-gray-500">Email:</span> {values.email}</p>
+                      <p><span className="text-gray-500">Password:</span> ••••••••</p>
+                    </div>
+                  </motion.div>
+                  
+                  <motion.div variants={itemVariants} className="flex items-center">
+                    <input
+                      id="terms"
+                      name="terms"
+                      type="checkbox"
+                      className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 rounded"
+                    />
+                    <label htmlFor="terms" className="ml-2 block text-sm text-gray-700">
+                      I agree to the <span className="text-indigo-600 cursor-pointer">Terms of Service</span> and <span className="text-indigo-600 cursor-pointer">Privacy Policy</span>
+                    </label>
+                  </motion.div>
+
+                  <div className="flex space-x-4">
+                    <motion.button
+                      type="button"
+                      onClick={prevStep}
+                      whileHover={{ scale: 1.03 }}
+                      whileTap={{ scale: 0.97 }}
+                      className="w-1/3 py-3 px-4 bg-gray-200 text-gray-700 font-medium rounded-lg hover:bg-gray-300 transition-all duration-200 ease-in-out"
+                    >
+                      Back
+                    </motion.button>
+                    <motion.button
+                      type="submit"
+                      disabled={isLoading}
+                      whileHover={{ scale: 1.03 }}
+                      whileTap={{ scale: 0.97 }}
+                      className="w-2/3 py-3 px-4 bg-gradient-to-r from-indigo-600 to-purple-600 text-white font-medium rounded-lg shadow-md hover:shadow-lg transition-all duration-300 ease-in-out relative overflow-hidden"
+                    >
+                      {isLoading ? (
+                        <>
+                          <span className="opacity-0">Create Account</span>
+                          <div className="absolute inset-0 flex items-center justify-center">
+                            <svg className="animate-spin h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                              <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                              <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                            </svg>
+                          </div>
+                        </>
+                      ) : "Create Account"}
+                    </motion.button>
+                  </div>
+                </motion.div>
+              )}
+            </AnimatePresence>
+          </form>
         </div>
-    </div>
+        <motion.div variants={itemVariants} className="px-8 py-4 bg-gray-50 border-t border-gray-100">
+          <p className="text-sm text-center text-gray-600">
+            Already have an account?{" "}
+            <Link to="/login" className="font-semibold text-indigo-600 hover:text-indigo-500 transition-colors">
+              Sign in
+            </Link>
+          </p>
+        </motion.div>
+      </div>
+    </motion.div>
   )
 }
 
