@@ -1,46 +1,203 @@
-import {Link} from 'react-router-dom'
+import { Link } from 'react-router-dom'
 import React, { useState } from 'react'
 import Validation from './LoginValidation';
+import { motion } from 'framer-motion';
 
+function Login() {
+  const [values, setValues] = useState({
+    email: '',
+    password: ''
+  })
+  const [errors, setErrors] = useState({})
+  const [isLoading, setIsLoading] = useState(false)
 
-function Login()  {
-      const[values , setValues] = useState({
-        email:'',
-        password:''
-      })
-      const [errors , setErrors] = useState({})
-      const handleInput = (event) => {
-        setValues({...values,[event.target.name]:event.target.value})
+  const handleInput = (event) => {
+    setValues({ ...values, [event.target.name]: event.target.value })
+  }
+
+  const handleSubmit = (event) => {
+    event.preventDefault()
+    const validationErrors = Validation(values)
+    setErrors(validationErrors)
+    
+    // Simulate login if no errors
+    if (Object.keys(validationErrors).length === 0) {
+      setIsLoading(true)
+      
+      // Simulate API call
+      setTimeout(() => {
+        setIsLoading(false)
+        // Here you would handle successful login
+      }, 1500)
+    }
+  }
+
+  // Animation variants
+  const containerVariants = {
+    hidden: { opacity: 0, y: 20 },
+    visible: { 
+      opacity: 1, 
+      y: 0,
+      transition: { 
+        duration: 0.5,
+        when: "beforeChildren",
+        staggerChildren: 0.1
       }
-      const handleSubmit = (event) => {
-        event.preventDefault()
-        setErrors(Validation(values));
-      }
+    },
+    exit: { 
+      opacity: 0, 
+      y: -20,
+      transition: { duration: 0.3 }
+    }
+  }
+
+  const itemVariants = {
+    hidden: { opacity: 0, y: 10 },
+    visible: { opacity: 1, y: 0 }
+  }
 
   return (
-    <div className='d-flex justify-content-center align-items-center bg-primary vh-100'>
-        <div className='bg-white p5 rounded w-50'>
-            <h2>Sign-In</h2>
-            <form action="" onSubmit={handleSubmit}>
-                 <div className='mb-3'>
-                   <label htmlFor='email'><strong>Email</strong></label>
-                   <input type="email" placeholder='Enter Email' name='email'  
-                   onChange={handleInput} className='form-control rounded-0'/>
-                   {errors.email && <span className='text-danger'> {errors.email}</span>}
-                 </div>
-                 <div className='mb=3'>
-                    <label htmlFor='password'><strong>Password</strong></label>
-                    <input type="password" placeholder='Enter password' name='password' 
-                    onChange={handleInput} className='form-control rounded-0'/>
-                    {errors.password && <span className='text-danger'> {errors.password}</span>}
+    <motion.div 
+      variants={containerVariants}
+      initial="hidden"
+      animate="visible"
+      exit="exit"
+      className="w-full max-w-md relative z-10"
+    >
+      {/* Decorative elements */}
+      <div className="absolute -top-10 -left-10 w-32 h-32 bg-indigo-500 rounded-full mix-blend-multiply filter blur-xl opacity-20 animate-float"></div>
+      <div className="absolute -bottom-10 -right-10 w-32 h-32 bg-purple-500 rounded-full mix-blend-multiply filter blur-xl opacity-20 animate-float" style={{ animationDelay: "1s" }}></div>
+      
+      <div className="bg-white shadow-2xl rounded-3xl overflow-hidden backdrop-blur-lg bg-opacity-95 border border-gray-100 relative z-10">
+        <div className="px-10 py-8">
+          <motion.div variants={itemVariants} className="text-center mb-8">
+            <h2 className="text-3xl font-bold mb-2 text-gray-800 font-display">Welcome Back</h2>
+            <p className="text-gray-500">Enter your credentials to access your account</p>
+          </motion.div>
+          
+          {/* Social login buttons */}
+          <motion.div variants={itemVariants} className="flex space-x-4 mb-6">
+            <button className="flex-1 py-2 px-4 bg-white border border-gray-300 rounded-lg hover:bg-gray-50 transition-colors flex items-center justify-center">
+              <svg className="w-5 h-5 mr-2" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+                <path d="M22.56 12.25c0-.78-.07-1.53-.2-2.25H12v4.26h5.92c-.26 1.37-1.04 2.53-2.21 3.31v2.77h3.57c2.08-1.92 3.28-4.74 3.28-8.09z" fill="#4285F4"/>
+                <path d="M12 23c2.97 0 5.46-.98 7.28-2.66l-3.57-2.77c-.98.66-2.23 1.06-3.71 1.06-2.86 0-5.29-1.93-6.16-4.53H2.18v2.84C3.99 20.53 7.7 23 12 23z" fill="#34A853"/>
+                <path d="M5.84 14.09c-.22-.66-.35-1.36-.35-2.09s.13-1.43.35-2.09V7.07H2.18C1.43 8.55 1 10.22 1 12s.43 3.45 1.18 4.93l2.85-2.22.81-.62z" fill="#FBBC05"/>
+                <path d="M12 5.38c1.62 0 3.06.56 4.21 1.64l3.15-3.15C17.45 2.09 14.97 1 12 1 7.7 1 3.99 3.47 2.18 7.07l3.66 2.84c.87-2.6 3.3-4.53 6.16-4.53z" fill="#EA4335"/>
+              </svg>
+              Google
+            </button>
+            <button className="flex-1 py-2 px-4 bg-[#1877F2] text-white rounded-lg hover:bg-[#0c65d8] transition-colors flex items-center justify-center">
+              <svg className="w-5 h-5 mr-2" fill="white" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24">
+                <path d="M24 12.073c0-6.627-5.373-12-12-12s-12 5.373-12 12c0 5.99 4.388 10.954 10.125 11.854v-8.385H7.078v-3.47h3.047V9.43c0-3.007 1.792-4.669 4.533-4.669 1.312 0 2.686.235 2.686.235v2.953H15.83c-1.491 0-1.956.925-1.956 1.874v2.25h3.328l-.532 3.47h-2.796v8.385C19.612 23.027 24 18.062 24 12.073z"/>
+              </svg>
+              Facebook
+            </button>
+          </motion.div>
+          
+          <motion.div variants={itemVariants} className="relative flex items-center my-6">
+            <div className="flex-grow border-t border-gray-200"></div>
+            <span className="mx-4 text-sm text-gray-400">or continue with</span>
+            <div className="flex-grow border-t border-gray-200"></div>
+          </motion.div>
+          
+          <form onSubmit={handleSubmit} className="space-y-6">
+            <motion.div variants={itemVariants}>
+              <label htmlFor="email" className="block text-sm font-medium text-gray-700 mb-1">Email Address</label>
+              <div className="relative">
+                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                  <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 12a4 4 0 10-8 0 4 4 0 008 0zm0 0v1.5a2.5 2.5 0 005 0V12a9 9 0 10-9 9m4.5-1.206a8.959 8.959 0 01-4.5 1.207" />
+                  </svg>
+                </div>
+                <motion.input
+                  type="email"
+                  id="email"
+                  name="email"
+                  placeholder="you@example.com"
+                  onChange={handleInput}
+                  whileFocus={{ scale: 1.01 }}
+                  transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                  className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                />
+              </div>
+              {errors.email && <p className="mt-1 text-sm text-red-600">{errors.email}</p>}
+            </motion.div>
+            
+            <motion.div variants={itemVariants}>
+              <div className="flex justify-between items-center mb-1">
+                <label htmlFor="password" className="block text-sm font-medium text-gray-700">Password</label>
+                <Link to="/forgot-password" className="text-xs text-indigo-600 hover:text-indigo-800 transition-colors">
+                  Forgot Password?
+                </Link>
+              </div>
+              <div className="relative">
+                <div className="absolute inset-y-0 left-0 pl-3 flex items-center pointer-events-none">
+                  <svg className="w-5 h-5 text-gray-400" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                    <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 15v2m-6 4h12a2 2 0 002-2v-6a2 2 0 00-2-2H6a2 2 0 00-2 2v6a2 2 0 002 2zm10-10V7a4 4 0 00-8 0v4h8z" />
+                  </svg>
+                </div>
+                <motion.input
+                  type="password"
+                  id="password"
+                  name="password"
+                  placeholder="••••••••"
+                  onChange={handleInput}
+                  whileFocus={{ scale: 1.01 }}
+                  transition={{ type: "spring", stiffness: 400, damping: 10 }}
+                  className="w-full pl-10 pr-4 py-2 border border-gray-300 rounded-lg focus:ring-2 focus:ring-indigo-500 focus:border-indigo-500 outline-none transition-all duration-200"
+                />
+              </div>
+              {errors.password && <p className="mt-1 text-sm text-red-600">{errors.password}</p>}
+            </motion.div>
 
-                 </div>
-                 <button type='submit'className='btn btn-success w-100 rounded-0'><strong>Login</strong></button>
-                 <p> You agree to our terms and conditions</p>
-                 <Link to="/signup" className='btn btn-default border w-100 bg-light rounded-0 text-decoration-none'>Create Account</Link>
-            </form>
+            <motion.div variants={itemVariants} className="flex items-center">
+              <input
+                id="remember-me"
+                name="remember-me"
+                type="checkbox"
+                className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 rounded"
+              />
+              <label htmlFor="remember-me" className="ml-2 block text-sm text-gray-700">
+                Remember me
+              </label>
+            </motion.div>
+
+            <motion.button
+              variants={itemVariants}
+              type="submit"
+              disabled={isLoading}
+              whileHover={{ scale: 1.03 }}
+              whileTap={{ scale: 0.97 }}
+              className={`w-full py-3 px-4 bg-gradient-to-r from-indigo-600 to-purple-600 text-white font-medium rounded-lg shadow-md hover:shadow-lg transition-all duration-300 ease-in-out relative overflow-hidden ${isLoading ? 'opacity-90' : ''}`}
+            >
+              {isLoading ? (
+                <>
+                  <span className="opacity-0">Sign In</span>
+                  <div className="absolute inset-0 flex items-center justify-center">
+                    <svg className="animate-spin h-5 w-5 text-white" xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24">
+                      <circle className="opacity-25" cx="12" cy="12" r="10" stroke="currentColor" strokeWidth="4"></circle>
+                      <path className="opacity-75" fill="currentColor" d="M4 12a8 8 0 018-8V0C5.373 0 0 5.373 0 12h4zm2 5.291A7.962 7.962 0 014 12H0c0 3.042 1.135 5.824 3 7.938l3-2.647z"></path>
+                    </svg>
+                  </div>
+                </>
+              ) : "Sign In"}
+            </motion.button>
+            
+            <motion.p variants={itemVariants} className="text-xs text-center text-gray-500 mt-4">
+              By continuing, you agree to our <span className="underline cursor-pointer">Terms of Service</span> and <span className="underline cursor-pointer">Privacy Policy</span>
+            </motion.p>
+          </form>
         </div>
-    </div>
+        <motion.div variants={itemVariants} className="px-8 py-4 bg-gray-50 border-t border-gray-100">
+          <p className="text-sm text-center text-gray-600">
+            Don't have an account?{" "}
+            <Link to="/signup" className="font-semibold text-indigo-600 hover:text-indigo-500 transition-colors">
+              Create an account
+            </Link>
+          </p>
+        </motion.div>
+      </div>
+    </motion.div>
   )
 }
 
